refactor: 미수정된 파일 발견으로 인한 추가적인 구조 리팩토링
