+++ src/pages/shelterInfo/components/VShelterInfo.tsx
@@ -2,37 +2,7 @@ import Pagination from 'commons/components/VPagenation';
 import ProfileCard from 'pages/profileList/components/ProfileCard';
 import { getCookie } from 'commons/cookie/cookie';
 import VShelterCard from './VShelterCard';
-
-export interface ShelterInfoProps {
-  name: string;
-  id: number;
-  adress: string;
-  call: string;
-}
-
-export interface PageNationProps {
-  setCurrentPage: (page: number) => void;
-  currentPage: number;
-  lastPage: number;
-  maxLength: number;
-}
-
-export interface ProfileProps {
-  map(
-    arg0: (item: any, index: number) => import('react/jsx-runtime').JSX.Element,
-  ): import('react').ReactNode;
-  profileImageUrl: string;
-  id: number;
-  name: string;
-  adoptionStatus: string;
-  age: number;
-}
-
-export interface ShelterInfoPageProps {
-  profileProps: ProfileProps;
-  shelterInfoProps: ShelterInfoProps;
-  pageNationProps: PageNationProps;
-}
+import { ShelterInfoPageProps } from '../shelterInfoType';
 
 const VShelterInfo = (props: ShelterInfoPageProps) => {
   const loginAccount = getCookie('accountInfo');
